requirements.txt
appdirs==1.4.2
Django==1.10.6
django-email-confirm-la==2.3.0
djangorestframework==3.5.4
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
